backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
pdfplumber
pydantic
python-multipart
requests
